+++ src/GSS.Authorization.OAuth2/AuthorizerOptions.cs
@@ -1,15 +1,19 @@
-using System;
+﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Net;
 
 namespace GSS.Authorization.OAuth2
 {
     public class AuthorizerOptions
     {
+        [Required]
         public Uri AccessTokenEndpoint { get; set; }
 
+        [Required]
         public string ClientId { get; set; }
 
+        [Required]
         public string ClientSecret { get; set; }
 
         public IEnumerable<string> Scopes { get; set; }